requirements.txt
Django==3.0.14
djangorestframework==3.13.1
ipython==8.4.0
python-dotenv==0.20.0
django-extensions==3.1.5
drf-yasg==1.20.0
